Always set both result values bc_num_split
In case of a split boundary beyond the length of the input value, the
second result argument was left unchanged. It should be a zero length
number or consistent semantics of this function.

While here add two asserts to make sure that the result values will
fit into the passed BcNum arrays.

Since this function is only called in two places in bc_num_k and all
arguments are initialized in such a way that no overflow occurs and
the second argument is initialized to zero in both cases, there is
no actual change, but the code is not robust in case bc_num_split was
ever used with less carefully prepared arguments.