Speed up the simple multiplication
The sum variable is known to have some head-room beyond the product of
two BcDig values.

Since it is known to be larger than 2 * BC_BASE_POW^2, the test for
overflow can be against BC_BASE_POW^2 in the loop, removing the div
and mod operations from nearly all iterations.

Only a single div and mod has to be performed after termination of
the loop to get the correct results.