Combine bc_num_a and bc_num_s in one function bc_num_as
The new code is only compiled in when building with -DUSE_SE_ADD, to
simplify comperative benchmarking.

Subtraction is implemented as addition of the negated second operand.
The operation is split into 2 steps:

1) Processing of low BcDig array elements present in only one
   operand (only if the rdx values are different).

2) Addition or subtraction of 2 values with same effective rdx
   values.

There are 4 cases for the first step, dependingg on the signs and
magnitudes of the 2 operands. The low end BcDig elements are either
copied unchanged or as their complement from the operand with higher
rdx to the corresponding array elements of the result. (These 4 caes
can be combined into 3 cases by operating on left and right operands
instead of a and b.)

The second step consists of addition of the two operands or the
subtraction of the smaller from the larger value, depending on whether
the signs of the operands do match or not.

The addition or subtraction of BcDig values is inlined in optimized
form by gcc and clang, leading to elimination of constant arguments.

This version is significantly shorter than the two functions it
replaces, if comments are not included in the line count. This is
due to less code replicated in both functions. The invocations in
bc_num_add and bc_num_sub are also simplified (no need for a function
pointer).

This code is significantly faster when compiling with gcc (in the
order of 2% for make test when built with gcc-9.1 and -O3 -flto),
but only slightly faster when built with clang.